[♻️Refactor]: 리팩토링 작업 중
- 함수 분리 및 중복코드 제거
