+++ src/lib/detail/calcuateTotalPrice.js
@@ -1,0 +1,10 @@
+export function calculateTotalPrice(
+  productQuantity,
+  floorDiscountPrice,
+  quantity,
+  totalPrice
+) {
+  productQuantity.textContent = quantity;
+  let total = floorDiscountPrice * quantity;
+  totalPrice.textContent = total.toLocaleString();
+}
