+++ src/lib/detail/index.js
@@ -1,4 +1,12 @@
 export * from './insert.js';
-export * from './fetchData.js';
+export * from './activeTab.js';
+export * from './accordion.js';
+export * from './updateCart.js';
 export * from './loginCheck.js';
+export * from './secretCheck.js';
+export * from './toggleContent.js';
+export * from './wishListChecked.js';
+export * from './reviewAndInquiry.js';
+export * from './accumulatePoints.js';
+export * from './calcuateTotalPrice.js';
 export * from './generateInfoSection.js';
