+++ src/lib/detail/insert.js
@@ -3,3 +3,8 @@ export function insertAfterBegin(node, renderTemplate) {
 
   node.insertAdjacentHTML('afterbegin', renderTemplate);
 }
+export function insertBeforeEnd(node, renderTemplate) {
+  node = document.querySelector(node);
+
+  node.insertAdjacentHTML('beforeend', renderTemplate);
+}
