+++ src/pages/productDetail/index.html
@@ -82,7 +82,9 @@
                 </p> -->
               </div>
               <!-- point -->
-              <p class="font-semibold text-primary text-label-medium">
+              <p
+                class="font-semibold text-primary text-label-medium login_point"
+              >
                 로그인 후, 적립 혜택이 제공됩니다.
               </p>
             </div>
@@ -201,7 +203,9 @@
                 >
                   적립
                 </span>
-                <span class="text-label-medium">로그인 후, 적립 혜택 제공</span>
+                <span class="text-label-medium login_point"
+                  >로그인 후, 적립 혜택 제공</span
+                >
               </p>
             </div>
